requirements.txt
Flask==3.1.1
Pillow==11.2.1
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
charset-normalizer==3.4.2
idna==3.10
certifi==2025.6.15 